% Generated by build script
\name{dryad_datasets}
\alias{dryad_datasets}
\title{List Dryad datasets}
\usage{
dryad_datasets()
}
\arguments{
\item{per_page}{Number of results per page (default 10, max 100)}
\item{page}{Page number (default 1)}
}
\value{tibble: id, doi, title, abstract, published_date, storage_size}
\description{List Dryad datasets}
