% Generated by build script
\name{dryad_files}
\alias{dryad_files}
\title{List files in a Dryad dataset}
\usage{
dryad_files()
}
\arguments{
\item{doi}{Dataset DOI (e.g. "doi:10.5061/dryad.1234")}
}
\value{tibble: path, size, mime_type, status, description}
\description{List files in a Dryad dataset}
