% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad.R
\name{dryad_datasets}
\alias{dryad_datasets}
\title{List Dryad datasets}
\usage{
dryad_datasets(per_page = 10, page = 1)
}
\arguments{
\item{per_page}{Number of results per page (default 10, max 100)}

\item{page}{Page number (default 1)}
}
\value{
tibble: id, doi, title, abstract, published_date, storage_size
}
\description{
List Dryad datasets
}
